Add "smart" JSON parsing of Data field
Adds "smart" parsing of the Data field into JSON, in that it tries to
parse it as JSON and falls back to []byte if there is any error in the
decoding.

Also factored the main function into a proper, top-level function, and
factored the "get shard iterator" logic into a separate function to
reduce the main func's complexity.